Modified the simmap_view.pt to use ++resources++ rather than harding coding a URL to the javascript, css, and images.  The javascript and css are still configured directly within the simmap_view.pt rather than through their repsective registries.
Modifed the unzip of GIS layers to convert the filename to lower case.  It
seems that having extensions in different cases (ie .shp and .DBF) caused
mapserver to bail.

Fixed get_mapserve method to call the plone.leamgroup.com instead of 
datacenter.leamgroup.com but this is still hardcoded and needs to 
get it from the simmap configuration.